+++ CHANGELOG.md
@@ -15,7 +15,8 @@
 - server: log all HTTP errors in remote schema calls as `remote-schema-error` with details
 - server: For BigQuery, make `global_select_limit` configuration optional with a default value of
 `1000`
-
+- console: add `reload all databases` checkbox to the metadata settings page
+- console: add schema sharing
 
 ## v2.0.1